IB/iser: Support up to 8MB data transfer in a single command
iser support up to 512KB data transfer in a single scsi
command. In order to support up to 8MB, iser needs to pre-allocate
larger memory regions and larger page vectors.

Given that a few target implementations don't support data transfers
of more than 512KB by default, we also introduce a module parameter
to determine the maximum number of 512B sectors in a single scsi
command. Users that are interested in 1MB transfers can change this
value given that the target supports 1MB transfers.

IO operations that consists of N pages will need a page vector
of size N+1 in case the first SG element contains an offset. Given
that some devices allocates memory regions in powers of 2, this
means that allocating a region with N+1 pages, will result in
region resources allocation of the next power of 2. Since we don't
want that to happen, in case we are in the limit of IO size supported
and the first SG element has an offset, we align the SG list using a
bounce buffer (which is semi OK given that this is not likely to
happen a lot).

Issue: 550273

Change-Id: I18f311bcdc619bdaf3560a74ba43d3cfeee4564d
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>