IB/hfi1: Add Infiniband dependency to Kconfig
The hfi1 driver depends on the infiniband core stack.
Building a kernel without the infiniband core but with the
hfi1 driver causes undefined reference errors to the ib_* functions.

Reported by 0-day build:
https://lists.01.org/pipermail/kbuild-all/2015-August/011216.html

This is a temporary patch and is only needed while the hfi1 driver is
in staging. Remove when moving driver back to drivers/infiniband.

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>